android-sdk/src/main/java/io/relayr/ble/service/LongWriteDataParser.java
package io.relayr.ble.service;

import java.nio.ByteBuffer;
import java.util.Arrays;

public class LongWriteDataParser {

    private final int CHUNK_OFFSET = 2;
    private final int CHUNK_FIRST_SIZE = 14;
    private final int CHUNK_DEFAULT_SIZE = 16;

    private boolean firstPackage = true;
    private int start = 0;
    private int end = 0;
    private byte[] data;

    public LongWriteDataParser(byte[] data) {
        this.data = data == null ? new byte[]{} : data;
    }

    public byte[] getData() {
        //if package is empty
        if (firstPackage && data.length == 0) {
            firstPackage = false;
            return new byte[]{0, 0, 0, 0};
        }
        //if there is no more data return empty array
        if (!firstPackage && end == data.length) return new byte[]{};

        int chunkSize = CHUNK_DEFAULT_SIZE;
        int chunkOffset = CHUNK_OFFSET;

        if (start == 0) {
            chunkSize = data.length < CHUNK_FIRST_SIZE ? data.length : CHUNK_FIRST_SIZE;
            chunkOffset = 4;
            end = chunkSize;
        } else {
            end = start + chunkSize > data.length ? data.length : start + chunkSize;
            chunkSize = end - start;
        }

        byte[] payload = new byte[chunkSize + chunkOffset];

        byte[] offset = ByteBuffer.allocate(2).putShort((short) (start << 8)).array();
        System.arraycopy(offset, 0, payload, 0, offset.length);

        if (start == 0) {
            byte[] length = ByteBuffer.allocate(2).putShort((short) (data.length << 8)).array();
            System.arraycopy(length, 0, payload, 2, length.length);
        }

        byte[] chunk = Arrays.copyOfRange(data, start, end);
        System.arraycopy(chunk, 0, payload, chunkOffset, chunk.length);

        start += chunkSize;
        firstPackage = false;

        return payload;
    }
}
